serverless/cloud-run/agents/adk-with-gpu/requirements.txt
pulumi>=3.203.0,<4.0.0
pulumi-gcp>=9.3.0,<9.5.0
pulumi-docker>=4.9.0,<5.0.0

pulumi-docker-build>=0.0.15,<1.0.0
google-auth>=2.41.1,<3.0.0
google-adk>=1.17.0,<2.0.0
fastapi>=0.119.1
locust>=2.42.0,<2.45.0
python-dotenv>=1.1.1,<1.5.0

